+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/screen/history/AndroidHistory.kt
@@ -1,4 +1,4 @@
-package co.electriccoin.zcash.ui.history
+package co.electriccoin.zcash.ui.screen.history
 
 import androidx.activity.ComponentActivity
 import androidx.activity.viewModels
@@ -6,7 +6,7 @@ import androidx.compose.runtime.Composable
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import cash.z.ecc.android.sdk.Synchronizer
 import co.electriccoin.zcash.ui.MainActivity
-import co.electriccoin.zcash.ui.history.view.History
+import co.electriccoin.zcash.ui.screen.history.view.History
 import co.electriccoin.zcash.ui.screen.home.viewmodel.WalletViewModel
 import kotlinx.collections.immutable.toImmutableList
